backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
openai==1.3.5
python-multipart==0.0.6
groq==0.11.0
httpx==0.25.0
